Checkout tagging dtype fix from other branch

+++ experiments/tagging/experiment.py
@@ -317,7 +317,7 @@ class TaggingExperiment(BaseExperiment):
 
     def _batch_loss(self, batch):
         y_pred, label, tracker, _ = self._get_ypred_and_label(batch)
-        loss = self.loss(y_pred, label.to(torch.float32))
+        loss = self.loss(y_pred, label)
         assert torch.isfinite(loss).all()
 
         metrics = tracker
