+++ experiments/tagging/topxlexperiment.py
@@ -141,4 +141,4 @@ class TopXLTaggingExperiment(TaggingExperiment):
         embedding = embed_tagging_data(fourmomenta, scalars, ptr, self.cfg.data)
         y_pred, tracker, lframes = self.model(embedding)
         y_pred = y_pred[:, 0]
-        return y_pred, label, tracker, lframes
+        return y_pred, label.to(y_pred.dtype), tracker, lframes
